+++ src/UI/InventoryMenu.h
@@ -1,8 +1,14 @@
 #pragma once
 
+namespace RE {
+class InventoryMenu;
+}
+
 namespace UI::InventoryMenu {
 [[nodiscard]] bool LastOpenedMenuUsesVanillaBottomBar();
-void QueueOpenMenuRingRowRefresh();
-void QueueRingRowRefresh();
-void QueueRefreshAfterVanillaRingSlotMove();
+void OnClosed();
+void OnShown(RE::InventoryMenu& a_inventoryMenu);
+void OnInventoryUpdateProcessed(RE::InventoryMenu& a_inventoryMenu);
+void RefreshAfterNextInventoryUpdate();
+void QueueInventoryListRefresh();
 }